Remove obsolete `wrangler` subdirectory. Move TravisCI from tests to its own directory.
