+++ tests/dask/test_base.py
@@ -8,7 +8,7 @@ import pytest
 pytestmark = pytest.mark.dask  # noqa: E402
 dask = pytest.importorskip("dask")  # noqa: E402
 
-from pywrangler.wranglers.dask.base import DaskWrangler
+from pywrangler.dask.base import DaskWrangler
 
 
 def test_dask_base_wrangler_engine():
