+++ tests/spark/test_base.py
@@ -9,7 +9,7 @@ import pandas as pd
 pytestmark = pytest.mark.pyspark  # noqa: E402
 pyspark = pytest.importorskip("pyspark")  # noqa: E402
 
-from pywrangler.wranglers.spark.base import SparkWrangler
+from pywrangler.spark.base import SparkWrangler
 
 
 def test_spark_base_wrangler_engine():
