+++ travisci/coveralls_master.sh
